requirements.txt
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-toolbelt==0.0.1
gunicorn==19.8.1
heroku==0.1.4
idna==2.6
psycopg2==2.7.4
python-dateutil==1.5
python-decouple==3.1
pytz==2018.4
requests==2.18.4
static3==0.7.0
urllib3==1.22
whitenoise==3.3.1
